requirements.txt
aiogram==3.7.0
httpx==0.27.0
chromadb==0.5.3
sentence-transformers==3.0.1
APScheduler==3.10.4
python-dotenv==1.0.1
SQLAlchemy==2.0.31
aiosqlite==0.20.0
PyMuPDF==1.24.9
python-docx==1.1.2
